consumer/requirements.txt
dnspython==2.7.0
kafka-python==2.0.6
pymongo==4.11.2
